# TwitterCloneApi/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Twitter Clone API
  description: API for the Twitter Clone application
  version: 0.0.1
servers:
  - url: http://localhost:3000
    description: Local server
  - url: https://twitter-clone-api.herokuapp.com
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/auth/register:
    post:
      summary: Register
      description: Register to the application
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  examples:
                    - joblipat
                password:
                  type: string
                  examples:
                    - password
      responses:
        200:
          description: Register successful
        400:
          description: Bad Request
  /api/v1/auth/login:
    post:
      summary: Login
      description: Login to the application
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  examples:
                    - joblipat
                password:
                  type: string
                  examples:
                    - password
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT token
                    examples:
                      - eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        400:
          description: Bad Request
...